+++ src/main/java/rbasamoyai/createbigcannons/munitions/autocannon/ammo_container/AutocannonAmmoContainerBlockEntity.java
@@ -111,9 +111,11 @@ public class AutocannonAmmoContainerBlockEntity extends BlockEntity implements I
 		this.name = name;
 	}
 
-	protected Component getDefaultName() {
-		return Component.translatable(CBCBlocks.AUTOCANNON_AMMO_CONTAINER.get().getDescriptionId());
-	}
+    protected Component getDefaultName() {
+        return Component.translatable(this.isCreativeContainer()
+            ? CBCBlocks.CREATIVE_AUTOCANNON_AMMO_CONTAINER.get().getDescriptionId()
+            : CBCBlocks.AUTOCANNON_AMMO_CONTAINER.get().getDescriptionId());
+    }
 
 	@Override
 	public Component getName() {