+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Fixed:
 - Fixed missing FramedBlocks block armor properties compatibilities
 - Fixed honey bucket and chocolate buckets not showing up in creative tab
 - Fixed melting basin ponder category and pump orientation
+- Fixed sliding breech and quick-firing breech breechblock rendering
 
 ## [5.10.2] - 2026-01-17